+++ src/generic-provider-views/index.js
@@ -34,6 +34,8 @@ const Utils = require('../core/Utils')
  *    @return {String} unique request path of the item when making calls to uppy server
  * getItemModifiedDate
  *    @return {object} or {String} date of when last the item was modified
+ * getItemThumbnailUrl
+ *    @return {String}
  */
 module.exports = class View {
   /**
@@ -49,7 +51,6 @@ module.exports = class View {
     this.filterQuery = this.filterQuery.bind(this)
     this.getFolder = this.getFolder.bind(this)
     this.getNextFolder = this.getNextFolder.bind(this)
-    // this.handleRowClick = this.handleRowClick.bind(this)
     this.logout = this.logout.bind(this)
     this.handleAuth = this.handleAuth.bind(this)
     this.handleDemoAuth = this.handleDemoAuth.bind(this)
@@ -57,6 +58,7 @@ module.exports = class View {
     this.sortByDate = this.sortByDate.bind(this)
     this.isActiveRow = this.isActiveRow.bind(this)
     this.handleError = this.handleError.bind(this)
+    this.handleScroll = this.handleScroll.bind(this)
 
     // Visual
     this.render = this.render.bind(this)
@@ -70,6 +72,18 @@ module.exports = class View {
     const {state} = this.plugin.core
 
     this.plugin.core.setState({[stateId]: Object.assign({}, state[stateId], newState)})
+  }
+
+  _updateFilesAndFolders (res, files, folders) {
+    this.plugin.getItemSubList(res).forEach((item) => {
+      if (this.plugin.isFolder(item)) {
+        folders.push(item)
+      } else {
+        files.push(item)
+      }
+    })
+
+    this.updateState({ folders, files })
   }
 
   /**
@@ -94,18 +108,8 @@ module.exports = class View {
           updatedDirectories = state.directories.concat([{id, title: name || this.plugin.getItemName(res)}])
         }
 
-        this.plugin.getItemSubList(res).forEach((item) => {
-          if (this.plugin.isFolder(item)) {
-            folders.push(item)
-          } else {
-            files.push(item)
-          }
-        })
-
-        let data = {folders, files, directories: updatedDirectories}
-        this.updateState(data)
-
-        return data
+        this._updateFilesAndFolders(res, files, folders)
+        this.updateState({ directories: updatedDirectories })
       },
       this.handleError)
   }
@@ -124,7 +128,7 @@ module.exports = class View {
     const tagFile = {
       source: this.plugin.id,
       data: this.plugin.getItemData(file),
-      name: this.plugin.getItemName(file),
+      name: this.plugin.getItemName(file) || this.plugin.getItemId(file),
       type: this.plugin.getMimeType(file),
       isRemote: true,
       body: {
@@ -141,7 +145,7 @@ module.exports = class View {
 
     Utils.getFileType(tagFile).then(fileType => {
       if (Utils.isPreviewSupported(fileType[1])) {
-        tagFile.preview = `${this.plugin.opts.host}/${this.Provider.id}/thumbnail/${this.plugin.getItemRequestPath(file)}`
+        tagFile.preview = this.plugin.getItemThumbnailUrl(file)
       }
       this.plugin.core.log('Adding remote file')
       this.plugin.core.emitter.emit('core:file-add', tagFile)
@@ -166,19 +170,6 @@ module.exports = class View {
         }
       }).catch(this.handleError)
   }
-
-  /**
-   * Used to set active file/folder.
-   * @param  {Object} file   Active file/folder
-   */
-  // handleRowClick (file) {
-  //   const state = this.plugin.core.getState()[this.plugin.stateId]
-  //   const newState = Object.assign({}, state, {
-  //     activeRow: this.plugin.getItemId(file)
-  //   })
-
-  //   this.updateState(newState)
-  // }
 
   filterQuery (e) {
     const state = this.plugin.core.getState()[this.plugin.stateId]
@@ -297,17 +288,27 @@ module.exports = class View {
     this.updateState({ error })
   }
 
+  handleScroll (e) {
+    const scrollPos = e.target.scrollHeight - (e.target.scrollTop + e.target.offsetHeight)
+    const path = this.plugin.getNextPagePath ? this.plugin.getNextPagePath() : null
+
+    if (scrollPos < 50 && path && !this._isHandlingScroll) {
+      this.Provider.list(path)
+        .then((res) => {
+          const { files, folders } = this.plugin.core.getState()[this.plugin.stateId]
+          this._updateFilesAndFolders(res, files, folders)
+        }).catch(this.handleError)
+        .then(() => { this._isHandlingScroll = false }) // always called
+
+      this._isHandlingScroll = true
+    }
+  }
+
   // displays loader view while asynchronous request is being made.
   _loaderWrapper (promise, then, catch_) {
     promise
-      .then((result) => {
-        this.updateState({ loading: false })
-        then(result)
-      })
-      .catch((err) => {
-        this.updateState({ loading: false })
-        catch_(err)
-      })
+      .then(then).catch(catch_)
+      .then(() => this.updateState({ loading: false })) // always called.
     this.updateState({ loading: true })
   }
 
@@ -338,14 +339,14 @@ module.exports = class View {
       addFile: this.addFile,
       filterItems: this.filterItems,
       filterQuery: this.filterQuery,
-      handleRowClick: this.handleRowClick,
       sortByTitle: this.sortByTitle,
       sortByDate: this.sortByDate,
       logout: this.logout,
       demo: this.plugin.opts.demo,
       isActiveRow: this.isActiveRow,
       getItemName: this.plugin.getItemName,
-      getItemIcon: this.plugin.getItemIcon
+      getItemIcon: this.plugin.getItemIcon,
+      handleScroll: this.handleScroll
     })
 
     return Browser(browserProps)